hello-world/05-multiplex
* removed the Runnable class (too complicated)
* extracted an upcounter class
* added a debug class (trace/debug/info/error)
* added an led flipper for board debugging
* added basic tests

This is where I go back to the basics and try to figure out
how c++ really works, and how to do basic stuff with timers
instead of the 'loop()' construct, as I need to increment
the digits once / second (1Hz) but need to refresh the LED
segments at something more like 1kHz.

Successes:
* using the timers is not hard
* strobing the LEDs works as expected
* added rudimentary tests, and they pass

Failures:
* the leds are lit, but the counter is not incrementing them
* the entire toolchain, from clion to the plugin to the ttyUSB is
like working with broken glass
* the code sometimes runs, sometimes does not run
* also unclear as to when the code actually updates on the device

Thoughts:
In order to get test code running locally, I'm going to have to
really invest in understanding cmake and the entire c++
ecosystem. I'd rather invest this time in learning rust and that
realm, instead.