hello-world/08 rust timers
Note: I was not able to get timers working the way I would like. I am
leaving this code here as a learning exercise, but after this
experience, I am convinced it is far far easier to use timers, but leave
interrupts disabled and do manual timer compares in the main loop.